+++ src/commands/copyTableName.ts
@@ -1,0 +1,11 @@
+import BaseCommand from "../common/baseCommand";
+import { TableNode } from "../tree/tableNode";
+import * as vscode from 'vscode';
+
+export class copyTableNameCommand extends BaseCommand {
+
+  async run(tableNode: TableNode) {
+    vscode.env.clipboard.writeText(tableNode.table);
+  }
+
+}