+++ src/commands/runSelectTop.ts
@@ -1,8 +1,8 @@
 import { selectTopCommand } from './selectTop';
 import { TableNode } from "../tree/tableNode";
 
-export class selectTop1000Command extends selectTopCommand {
+export class runSelectTopCommand extends selectTopCommand {
   async run(treeNode: TableNode) {
-    return super.run(treeNode, 1000);
+    return super.run(treeNode, 0, false, true);
   }
 }