+++ src/commands/selectTop1000WithNames.ts
@@ -1,8 +1,10 @@
 import { selectTopCommand } from './selectTop';
 import { TableNode } from "../tree/tableNode";
 
-export class selectTop1000Command extends selectTopCommand {
+export class selectTop1000WithNamesCommand extends selectTopCommand {
+
   async run(treeNode: TableNode) {
-    return super.run(treeNode, 1000);
+    return super.run(treeNode, 1000, true);
   }
+
 }